♻️ refactor : Integration Test용 레포지토리 추가
