+++ src/test/java/com/devcourse/checkmoi/domain/comment/service/CommentQueryServiceImplTest.java
@@ -13,64 +13,28 @@ import static com.devcourse.checkmoi.util.EntityGeneratorUtil.makeStudyMember;
 import static com.devcourse.checkmoi.util.EntityGeneratorUtil.makeUser;
 import static org.assertj.core.api.Assertions.assertThat;
 import com.devcourse.checkmoi.domain.book.model.Book;
-import com.devcourse.checkmoi.domain.book.repository.BookRepository;
 import com.devcourse.checkmoi.domain.comment.converter.CommentConverter;
 import com.devcourse.checkmoi.domain.comment.dto.CommentRequest.Search;
 import com.devcourse.checkmoi.domain.comment.dto.CommentResponse.CommentInfo;
-import com.devcourse.checkmoi.domain.comment.repository.CommentRepository;
 import com.devcourse.checkmoi.domain.post.model.Post;
-import com.devcourse.checkmoi.domain.post.repository.PostRepository;
 import com.devcourse.checkmoi.domain.study.model.Study;
-import com.devcourse.checkmoi.domain.study.repository.StudyMemberRepository;
-import com.devcourse.checkmoi.domain.study.repository.StudyRepository;
 import com.devcourse.checkmoi.domain.user.model.User;
-import com.devcourse.checkmoi.domain.user.repository.UserRepository;
+import com.devcourse.checkmoi.template.IntegrationTest;
 import java.util.List;
 import java.util.stream.Stream;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
-class CommentQueryServiceImplTest {
+class CommentQueryServiceImplTest extends IntegrationTest {
 
     @Autowired
     private CommentQueryServiceImpl commentQueryService;
 
     @Autowired
-    private CommentRepository commentRepository;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private StudyRepository studyRepository;
-
-    @Autowired
-    private BookRepository bookRepository;
-
-    @Autowired
-    private StudyMemberRepository studyMemberRepository;
-
-    @Autowired
-    private PostRepository postRepository;
-
-    @Autowired
     private CommentConverter commentConverter;
-
-    @AfterEach
-    void cleanUp() {
-        commentRepository.deleteAllInBatch();
-        postRepository.deleteAllInBatch();
-        studyMemberRepository.deleteAllInBatch();
-        studyRepository.deleteAllInBatch();
-        bookRepository.deleteAllInBatch();
-        userRepository.deleteAllInBatch();
-    }
 
     @Nested
     @DisplayName("작성된 글에 대한 댓글 목록 조회 #130")
